requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
black==22.12.0
click==8.1.3
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.3.0
mango-jwt==1.3.2
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.3
platformdirs==2.6.2
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==0.21.1
pytz==2022.7.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
asgiref==3.6.0
backoff==2.2.1
beautifulsoup4==4.11.2
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.3.0
idna==3.4
ipify==1.0.0
mango-jwt==1.3.2
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.3
platformdirs==2.6.2
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
